pyproject.toml
[project]
name = "madr"
version = "0.1.0"
description = ""
authors = [
    {name = "Samael Picoli",email = "samaelmunizpicoli27@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "freezegun (>=1.5.2,<2.0.0)"
]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.1"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev madr/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=madr -vv'
post_test = 'coverage html'

[tool.coverage.run]
concurrency = ['thread', 'greenlet']
